requirements.txt
https://github.com/AXERA-TECH/pyaxengine/releases/download/0.1.3.rc1/axengine-0.1.3-py3-none-any.whl
ml_dtypes==0.5.3
Pillow==12.0.0
Requests==2.32.5
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
sounddevice
librosa
kaldi_native_fbank
sentencepiece
onnxruntime
g2p_en
loguru
piper-tts
cn2an
pypinyin
jieba
